requirements.txt
streamlit==1.32.0
openai==0.28.1
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.3
numpy
soundfile
pyht==0.1.5